fix(tests): cast peak/requested lease bytes and ppm ratios for wider counters
Lease accounting counters widened to u128-sized peaks/requests in parts of
the stack; a few consumer tests still compared mixed widths and failed to
type-check or over-assert.

- fs-exec lease_battery: compare peak_bytes as u128 to requested_bytes
- fs-flagship-e2e production_scale: requested_bytes vs reservation_u64 as u128
- fs-lbm perf_model: cast RATIO_PPM to usize before sparse-tile arithmetic

No production logic change; tests only, restoring compile-compat and the
intended peak/occupancy assertions.

+++ crates/fs-exec/tests/lease_battery.rs
@@ -123,7 +123,10 @@ fn exact_accounting_is_deterministic_and_returns_to_zero() {
     assert!(a.requested_bytes > 0, "root metadata is charged");
     assert_eq!(a.used_bytes, 0, "every charge is released by run end");
     assert_eq!(a.refusals, 0);
-    assert_eq!(a.peak_bytes, a.requested_bytes, "one root charge, no churn");
+    assert_eq!(
+        a.peak_bytes as u128, a.requested_bytes,
+        "one root charge, no churn"
+    );
 }
 
 #[test]
@@ -288,7 +291,7 @@ fn varying_worker_counts_hold_the_lease_invariants() {
         );
         let receipt = lease.receipt();
         assert_eq!(receipt.used_bytes, 0, "workers={workers}");
-        assert!(receipt.peak_bytes <= receipt.requested_bytes);
+        assert!(receipt.peak_bytes as u128 <= receipt.requested_bytes);
         assert!(pool.arena_pool().stats().quiescent());
     }
 }
